requirements.txt
streamlit>=1.35.0
pandas>=2.2.0
requests>=2.31.0
qdrant-client>=1.9.0
sentence-transformers>=2.7.0
langchain-text-splitters>=0.2.0
python-dotenv>=1.0.1
